pyproject.toml
[project]
name = "bravebird-ephemeral-infra"
version = "0.1.0"
description = "Ephemeral environment orchestration for Computer Use tasks"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "boto3>=1.34.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-ulid>=2.2.0",
    "websockets>=12.0",
    "httpx>=0.26.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "moto[all]>=5.0.0",
    "localstack>=3.0.0",
]

[project.scripts]
cli = "src.cli.cli:app"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
